lambda-s3-trigger/requirements.txt
aws-cdk.core

aws-cdk.aws-s3
aws-cdk.aws-lambda
aws-cdk.aws_s3_notifications